Initial skeleton code
client: prototype to use tink libraries to perform processing on dek,
kek etc
api_server: provides two endpoints /login and /process
This basically returns a jwt to the client on successful login.
The client uses the jwt for subsequent request to send data to the
/process endpoint
process_server: this provides a single endpoint /process and processes
requests from the api_server. This program should run inside TEE to
perform decryption of the user sent data and processing the same

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>